Aggiunta alert "software non presente"

+++ src/main/java/corso/java/services/SoftwareServiceImpl.java
@@ -39,6 +39,10 @@ public class SoftwareServiceImpl implements SoftwareService{
 
 	@Override
 	public void eliminaSoftware(Long id) {
+		if (!softwareRepository.existsById(id)) {
+			throw new IllegalArgumentException("Software non presente in Database");
+			
+		}
 		softwareRepository.deleteById(id);
 		
 	}
